React Renderer: refactor engine logging out.
